Add a new people views & some methods to PersonDecorator
Inside of the 'people/show' view, I moved the wrapping div & its header
for associated enrollments inside of the partial. This is to match the
pattern established by the new '_teachers' and '_students' partial.

The reason for keeping the wrapping div & header inside of the partials
for teachers & students is because a person may not always have a
student or a teacher. It's cleaner this way. Though the individual
partials could be DRYed up in the future.